Move fallback language resources to separate file
