+++ src/@types/react-i18next.d.ts
@@ -1,6 +1,7 @@
 import "react-i18next";
 
-import { defaultNS, fallbackLngResources } from "../model/i18next.config";
+import { defaultNS } from "../model/i18next-config";
+import fallbackLngResources from "../model/i18next-fallback-resources";
 
 declare module "react-i18next" {
   interface CustomTypeOptions {
