+++ src/components/LanguageSwitcher.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
-import { supportedLngs } from "../model/i18next.config";
+import { supportedLngs } from "../model/i18next-config";
 
 const LanguageSwitcher = () => {
   const { i18n } = useTranslation();
