+++ src/model/i18next-config.ts
@@ -1,5 +1,3 @@
-import enTranslation from "../../static/locales/en/translation.json";
-
 // languages
 export const fallbackLng = "en";
 export const supportedLngs = [fallbackLng, "de"] as const;
@@ -7,8 +5,3 @@ export const supportedLngs = [fallbackLng, "de"] as const;
 // namespaces
 export const defaultNS = "translation";
 export const ns = [defaultNS] as const;
-
-// type
-export const fallbackLngResources = {
-  translation: enTranslation,
-};
